blog/blogs/requirements.txt
asgiref==3.8.1
bleach==6.2.0
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
django-extensions==3.2.3
django-summernote==0.8.20.0
djangorestframework==3.15.2
pillow==11.0.0
psycopg2==2.9.10
sqlparse==0.5.2
typing_extensions==4.12.2
webencodings==0.5.1
